--- a/src/pixie/masks.nim
+++ b/src/pixie/masks.nim
@@ -275,30 +275,24 @@ proc blur*(mask: Mask, radius: float32, outOfBounds: uint8 = 0) {.raises: [Pixie
     for x in 0 ..< mask.width:
       var value: uint32
       for xx in x - radius ..< min(x + radius, 0):
-        value += outOfBounds * kernel[xx - x + radius]
-
+        value += outOfBounds * kernel[xx - x + radius].uint32
       for xx in max(x - radius, 0) .. min(x + radius, mask.width - 1):
-        value += mask.getValueUnsafe(xx, y) * kernel[xx - x + radius]
-
+        value += mask.getValueUnsafe(xx, y) * kernel[xx - x + radius].uint32
       for xx in max(x - radius, mask.width) .. x + radius:
-        value += outOfBounds * kernel[xx - x + radius]
-
-      blurX.setValueUnsafe(y, x, (value div 1024 div 255).uint8)
+        value += outOfBounds * kernel[xx - x + radius].uint32
+      blurX.setValueUnsafe(y, x, (value div 256 div 255).uint8)
 
   # Blur in the Y direction and modify image.
   for y in 0 ..< mask.height:
     for x in 0 ..< mask.width:
       var value: uint32
       for yy in y - radius ..< min(y + radius, 0):
-        value += outOfBounds * kernel[yy - y + radius]
-
+        value += outOfBounds * kernel[yy - y + radius].uint32
       for yy in max(y - radius, 0) .. min(y + radius, mask.height - 1):
-        value += blurX.getValueUnsafe(yy, x) * kernel[yy - y + radius]
-
+        value += blurX.getValueUnsafe(yy, x) * kernel[yy - y + radius].uint32
       for yy in max(y - radius, mask.height) .. y + radius:
-        value += outOfBounds * kernel[yy - y + radius]
-
-      mask.setValueUnsafe(x, y, (value div 1024 div 255).uint8)
+        value += outOfBounds * kernel[yy - y + radius].uint32
+      mask.setValueUnsafe(x, y, (value div 256 div 255).uint8)
 
 when defined(release):
   {.pop.}