--- a/tests/benchmark_images_blur.nim
+++ b/tests/benchmark_images_blur.nim
@@ -41,13 +41,10 @@ proc blurSlower*(
       var values: array[4, uint32]
       for xx in x - radius ..< min(x + radius, 0):
         values += outOfBounds * kernel[xx - x + radius]
-
       for xx in max(x - radius, 0) .. min(x + radius, image.width - 1):
         values += image.getRgbaUnsafe(xx, y) * kernel[xx - x + radius]
-
       for xx in max(x - radius, image.width) .. x + radius:
         values += outOfBounds * kernel[xx - x + radius]
-
       blurX.setRgbaUnsafe(x, y, rgbx(values))
 
   # Blur in the Y direction.
@@ -56,13 +53,10 @@ proc blurSlower*(
       var values: array[4, uint32]
       for yy in y - radius ..< min(y + radius, 0):
         values += outOfBounds * kernel[yy - y + radius]
-
       for yy in max(y - radius, 0) .. min(y + radius, image.height - 1):
         values += blurX.getRgbaUnsafe(x, yy) * kernel[yy - y + radius]
-
       for yy in max(y - radius, image.height) .. y + radius:
         values += outOfBounds * kernel[yy - y + radius]
-
       image.setRgbaUnsafe(x, y, rgbx(values))
 
 let image = newImage(1920, 1080)