mask old:
`blur ............................. 165.291 ms    173.937 ms    ±6.383    x29`

mask new:
`blur ............................. 107.985 ms    115.985 ms    ±6.062    x43`

image old:
`blur ............................. 335.518 ms    349.339 ms    ±9.973    x15`

image new:
`blur ............................. 179.857 ms    188.839 ms    ±6.833    x27`